requirements.txt
requests>=2.32.3
dnspython>=2.7.0
colorama>=0.4.6
typing-extensions>=4.13.2
certifi>=2025.1.31
pymongo>=4.6.0
cryptography>=44.0.2
urllib3>=2.4.0
charset-normalizer>=3.4.1
idna>=3.10
jinja2>=3.1.6
plotly>=6.0.1
pandas>=2.2.3
PyJWT>=2.8.0
grpcio>=1.71.0
grpcio-tools>=1.71.0
grpcio-health-checking>=1.71.0
grpcio-reflection>=1.71.0
protobuf>=4.25.1
pyyaml>=6.0.1
numpy>=2.2.5 